feat: add updateComment functionality and added route; add one more check in deleteComment controller to check if user authorized

+++ src/controllers/comment.controller.js
@@ -32,6 +32,36 @@ const addComment = asyncHandler(async (req, res) => {
       );
 });
 
+const updateComment = asyncHandler(async (req, res) => {
+
+   const { commentId } = req.params;
+   const { content } = req.body;
+
+   validateMongoId(commentId);
+
+   if (isEmptyString(content))
+      throw new ApiError(400, "Comment content is required");
+
+   const comment = await Comment.findById(commentId);
+
+   if (!comment) throw new ApiError(404, "Comment not found");
+
+   if (!comment.owner.equals(req.user?._id))
+      throw new ApiError(403, "Not authorized to make changes on comment");
+
+   const updatedComment = await Comment.findByIdAndUpdate(
+      commentId,
+      { content },
+      { new: true }
+   );
+
+   return res
+      .status(200)
+      .json(
+         new ApiResponse(200, updatedComment, "Comment updated successfully")
+      );
+});
+
 const deleteComment = asyncHandler(async (req, res) => {
    const { commentId } = req.params;
 
@@ -41,6 +71,9 @@ const deleteComment = asyncHandler(async (req, res) => {
 
    if (!comment) throw new ApiError(404, "Comment not found");
 
+   if (!comment.owner.equals(req.user?._id))
+      throw new ApiError(403, "Not authorized to make changes on comment");
+
    await Comment.findByIdAndDelete(commentId);
 
    return res
@@ -48,4 +81,4 @@ const deleteComment = asyncHandler(async (req, res) => {
       .json(new ApiResponse(200, {}, "Comment deleted successfully"));
 });
 
-export { addComment, deleteComment };
+export { addComment, deleteComment, updateComment };
